working on documentation

+++ scripts/create-release.py
@@ -75,7 +75,10 @@ def make_docs():
 
     ex("make clean && make html", cwd=doc_build_dir)
 
-    remove_directory(doc_dest_dir, False)
+    if os.path.isdir(doc_dest_dir):
+        remove_directory(doc_dest_dir, False)
+    else:
+        os.makedirs(doc_dest_dir)
 
     items = set(os.listdir(doc_html_build)) ^ set(['.buildinfo', 'objects.inv'])
     for item in items:
@@ -141,3 +144,7 @@ if __name__ == '__main__':
         path = os.path.join(LIB_DIR, directory)
         if os.path.exists(path):
             remove_directory(path)
+
+    # clean up the doc build
+    doc_dir = os.path.join(LIB_DIR, 'docs')
+    remove_directory(doc_dir, remove_top=True)
